added logic to delete old unused files from cloudinary

+++ Backend/src/utils/cloudinary.js
@@ -8,36 +8,52 @@ dotenv.config();
 
 
  // Configuration
-    cloudinary.config({ 
-        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-        api_key: process.env.CLOUDINARY_API_KEY, 
-        api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
-    });
+cloudinary.config({ 
+    cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+    api_key: process.env.CLOUDINARY_API_KEY, 
+    api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
+});
 
 
-    const uploadOnCloudinary = async (localFilePath) => {
-        try {
-            if (!localFilePath) return null;
+const uploadOnCloudinary = async (
+        localFilePath,
+        folder = "misc" //if folder not defined
+    ) => {
+    try {
+        if (!localFilePath) return null;
 
-            const filePath = path.resolve(localFilePath);
+        const absolutePath = path.resolve(localFilePath);
 
-            const response = await cloudinary.uploader.upload(filePath, {
+        const result = await cloudinary.uploader.upload(absolutePath, {
             resource_type: "auto",
-            });
+            folder
+        });
 
-            // DELETE LOCAL FILE AFTER SUCCESS
-            fs.unlinkSync(filePath);
-
-            return response;
-
-        } catch (error) {
-            console.log("Cloudinary upload error:", error);
-
-            // DELETE LOCAL FILE IF FAILED
-            fs.unlinkSync(localFilePath);
-
-            return null;
+        // cleanup local file
+        if (fs.existsSync(absolutePath)) {
+        fs.unlinkSync(absolutePath);
         }
-    }
 
-    export {uploadOnCloudinary}
+        return {
+            url: result.secure_url,
+            public_id: result.public_id
+        };
+
+    } catch (error) {
+        console.error("Cloudinary upload error:", error);
+
+        if (fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
+        }
+
+        return null;
+    }
+};
+
+    const deleteFromCloudinary = async (publicId) => {
+    if (!publicId) return;
+
+    await cloudinary.uploader.destroy(publicId);
+    };
+
+    export {uploadOnCloudinary, deleteFromCloudinary}
